--- agent-memory-viewer/frontend/src/components/memory-views/PostedMemory.tsx
@@ -1,25 +1,0 @@
-import { PostedMemory } from '../../types';
-import { 
-    MemoryStatus,
-    AdditionalInfo,
-    Signature,
-    AgentVersion
-} from './shared/CommonSections';
-
-interface Props {
-    memory: PostedMemory;
-}
-
-export function PostedMemoryView({ memory }: Props) {
-    return (
-        <>
-            <MemoryStatus type={memory.type} />
-            <AdditionalInfo 
-                previousCid={memory.previousCid} 
-                timestamp={memory.timestamp} 
-            />
-            <Signature value={memory.signature} />
-            <AgentVersion version={memory.agentVersion} />
-        </>
-    );
-}
